renamed vocabs folder to just vocab
