+++ src/SupportedDataTypes.test.js
@@ -20,9 +20,7 @@ describe("Supported Data Type", () => {
         artifactGeneratorVersion: "^7.8.9",
         generatedTimestamp: "1999/12/31 23:59",
 
-        inputResources: [
-          "./test/resources/vocabs/special-case-owl-snippet.ttl",
-        ],
+        inputResources: ["./test/resources/vocab/special-case-owl-snippet.ttl"],
         outputDirectory,
         artifactVersion: "1.0.0",
         moduleNamePrefix: "generated-vocab-",
@@ -77,7 +75,7 @@ describe("Supported Data Type", () => {
         generatedTimestamp: "1999/12/31 23:59",
 
         inputResources: [
-          "./test/resources/vocabs/special-case-http-snippet.ttl",
+          "./test/resources/vocab/special-case-http-snippet.ttl",
         ],
         outputDirectory,
         artifactVersion: "1.0.0",
@@ -135,7 +133,7 @@ describe("Supported Data Type", () => {
     const generator = new VocabGenerator(
       {
         inputResources: [
-          "./test/resources/vocabs/supported-data-types-invalid-constant-iri.ttl",
+          "./test/resources/vocab/supported-data-types-invalid-constant-iri.ttl",
         ],
         outputDirectory,
       },
@@ -167,7 +165,7 @@ describe("Supported Data Type", () => {
     const generator = new VocabGenerator(
       {
         inputResources: [
-          "./test/resources/vocabs/supported-data-types-too-many-constant-iri-values.ttl",
+          "./test/resources/vocab/supported-data-types-too-many-constant-iri-values.ttl",
         ],
         outputDirectory,
       },
@@ -201,7 +199,7 @@ describe("Supported Data Type", () => {
     const generator = new VocabGenerator(
       {
         inputResources: [
-          "./test/resources/vocabs/supported-data-types-too-many-constant-string-values.ttl",
+          "./test/resources/vocab/supported-data-types-too-many-constant-string-values.ttl",
         ],
         outputDirectory,
       },
@@ -235,7 +233,7 @@ describe("Supported Data Type", () => {
     const generator = new VocabGenerator(
       {
         inputResources: [
-          "./test/resources/vocabs/supported-data-types-many-constant-string-values-but-different-languages.ttl",
+          "./test/resources/vocab/supported-data-types-many-constant-string-values-but-different-languages.ttl",
         ],
         outputDirectory,
       },
@@ -272,7 +270,7 @@ describe("Supported Data Type", () => {
         artifactGeneratorVersion: "^7.8.9",
         generatedTimestamp: "1999/12/31 23:59",
 
-        inputResources: ["./test/resources/vocabs/supported-data-types.ttl"],
+        inputResources: ["./test/resources/vocab/supported-data-types.ttl"],
         outputDirectory,
         artifactVersion: "1.0.0",
         moduleNamePrefix: "generated-vocab-",
