+++ test/resources/backwardCompatibility/vocab-list_no-packaging.yml
@@ -80,9 +80,9 @@ artifactToGenerate:
 vocabList:
   - descriptionFallback: Snippet of Schema.org from Google, Microsoft, Yahoo and Yandex
     inputResources:
-      - ../vocabs/schema-snippet.ttl
+      - ../vocab/schema-snippet.ttl
 
   - descriptionFallback: Vocab for testing predicate types...
     nameAndPrefixOverride: override-name
     inputResources:
-      - ../vocabs/supported-data-types.ttl
+      - ../vocab/supported-data-types.ttl
