+++ test/resources/packaging/vocab-list-no-packaging-templates.yml
@@ -26,4 +26,4 @@ artifactToGenerate:
 vocabList:
   - descriptionFallback: Snippet of Schema.org from Google, Microsoft, Yahoo and Yandex
     inputResources:
-      - ../vocabs/schema-snippet.ttl
+      - ../vocab/schema-snippet.ttl
