+++ test/resources/packaging/vocab-list-unpublish-command.yml
@@ -38,4 +38,4 @@ artifactToGenerate:
 vocabList:
   - descriptionFallback: Snippet of Schema.org from Google, Microsoft, Yahoo and Yandex
     inputResources:
-      - ../vocabs/schema-snippet.ttl
+      - ../vocab/schema-snippet.ttl
