+++ test/resources/vocab/isDefinedBy.ttl
