+++ test/resources/vocab/schema-inrupt-ext-no-author.ttl
