+++ test/resources/vocab/schema-inrupt-ext.ttl
