+++ test/resources/vocab/schema-snippet.ttl
