+++ test/resources/vocab/sosa.ttl
