+++ test/resources/vocab/special-case-http-snippet.ttl
