+++ test/resources/vocab/special-case-owl-snippet.ttl
