+++ test/resources/vocab/supported-data-types-invalid-constant-iri.ttl
