+++ test/resources/vocab/supported-data-types-many-constant-string-values-but-different-languages.ttl
