+++ test/resources/vocab/supported-data-types-no-author.ttl
