+++ test/resources/vocab/supported-data-types-too-many-constant-iri-values.ttl
