+++ test/resources/vocab/supported-data-types-too-many-constant-string-values.ttl
