+++ test/resources/vocab/supported-data-types.ttl
