+++ test/resources/vocab/term-with-label-but-no-comment.ttl
