+++ test/resources/vocab/vocab-list-missing-artifact-name.yml
